pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "payflow"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "pydantic[email]==2.8.0",
    "pydantic-settings==2.4.0",

    # Database
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.29.0",
    "alembic==1.13.2",

    # Cache & Queue
    "redis[hiredis]==5.0.8",
    "aiokafka==0.11.0",

    # Background tasks
    "celery[redis]==5.4.0",

    # HTTP client
    "httpx==0.27.0",
    "clickhouse-driver",

    # Observability
    "prometheus-client==0.21.0",
    "structlog==24.4.0",
    "sentry-sdk[fastapi]==2.13.0",

    # Utils
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
    "testcontainers[postgres,kafka,redis]==4.7.2",
    "psycopg2-binary==2.9.9",
    "ruff==0.6.3",
    "mypy==1.11.2",
    "locust==2.31.3",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
# E501: длинные русские поясняющие комментарии часто >88 символов; длину кода держим через ruff format.
ignore = ["B008", "E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
disable_error_code = ["call-arg", "misc", "no-redef"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# Порог 80% по всему пакету app (unit + integration).
[tool.coverage.run]
source = ["app"]

[tool.hatch.build.targets.wheel]
packages = ["app"]